requirements.txt
bs4
pymysql
loguru
python-dotenv
inflect